+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Strophe.js Change Log
 
-## Version 1.2.15 - (Unreleased)
+## Version 1.2.15 - (2018-05-21)
 * #259 XML element should be sent to xmlOutput
 * #266 Support Browserify/CommonJS. `require('strophe.js/src/wrapper')`
-* #288 Strophe logs fatal errors by default.
+* #296 Remove error handler from old websocket before closing
+* #271 SASL X-OAUTH2 authentication mechanism implemented 
+* #288 Strophe now logs fatal errors by default.
+* Run tests with headless Chromium instead of Phantomjs
 
 ## Version 1.2.14 - 2017-06-15
 * #231 SASL OAuth Bearer authentication should not require a JID node, when a user identifer